Add ExecuteSubsystemAsync for subsystem based proc
Adds the ExecuteSubsystemAsync call which can create an SSH channel from
a subsystem and return the RemoteProcess interface. This is used when
the SSH server defines custom subsystems in its sshd_config file that
allows a client to create a channel using a pre-defined subsystem name
rather than through an explicit command line argument.