[Breaking] Drop support for RSA PKCS#1 PEM keys in favor of the OpenSSH format.
As described in the README, this library embraces OpenSSH format when possible.

This drops support for the RSA PKCS#1 PEM ('RSA PRIVATE KEY') format which
was supported because the library author was using this format for a key.

If you are using a key in this format, it can be converted by running:

   ssh-keygen -p -f <key_file>

The OpenSSH format has been supported in OpenSSH since 2014 and became
the default format used by 'ssh-keygen' in 2018.